+++ CHANGELOG.md
@@ -215,6 +215,7 @@
 - Fix a bug when editing the root document @mpalomaki
 - Add position in parent to export script @robgietema
 - Return username in user endpoints @robgietema
+- Fix bug when uploading images and files @robgietema
 
 ### Internal